:sparkles: completed user and medical history controller

+++ backend/src/main/java/com/hms/repo/AddressRepo.java
@@ -2,8 +2,8 @@ package com.hms.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.hms.payload.Address;
+import com.hms.model.Address;
 
-public interface AddressRepo extends JpaRepository<Address, Long> {
+public interface AddressRepo extends JpaRepository<Address, Integer> {
 
 }
