+++ backend/src/main/java/com/hms/repo/PatientRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.hms.model.Patient;
 import com.hms.model.User;
 
-public interface PatientRepo extends JpaRepository<Patient, Long> {
+public interface PatientRepo extends JpaRepository<Patient, Integer> {
     Patient findByUser(User user);
 }
